Add support to cancel published jobs
This patch allows a ProcessBlock to cancel jobs it published
to child ProcessBlocks. Children will send the jobs they manage to
their parent until they all reach the ProcessBlock that emitted the
cancel event. From there the canceler can reprocess the jobs to its
liking.

It is interesting to note that brother ProcessBlocks cannot yet share
the jobs they choose to reschedule. This limits the degree to which
jobs that are canceled are properly balanced across blocks of a same
level.

This feature is important to be able to define dynamic controller on
a pipeline: without it, it is impossible to reformat a job that does
not fit an execution policy anymore.

There is currently no mechanism to cancel a running job. There will
not be any at the ProcessBlock level, as this class is intended to
be as generic as possible. However, Worker should and will support
cancelling a running job in the near future.

Change-Id: Iaa3092cdf46f32996f1c28d8f14c557a59396d21
Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/2584
Tested-by: Jenkins s8open_nr
Reviewed-by: Henri Doreau <henri.doreau@cea.fr>